+++ src/content/container.ts
@@ -3,10 +3,13 @@ import browser, { type Browser } from 'webextension-polyfill'
 
 import { createLogger, Logger } from '@/shared/logger'
 import { ContentScript } from './services/contentScript'
+import { MonetizationTagManager } from './services/monetizationTagManager'
 
 interface Cradle {
   logger: Logger
   browser: Browser
+  document: Document
+  monetizationTagManager: MonetizationTagManager
   contentScript: ContentScript
 }
 
@@ -20,6 +23,8 @@ export const configureContainer = () => {
   container.register({
     logger: asValue(logger),
     browser: asValue(browser),
+    document: asValue(document),
+    monetizationTagManager: asClass(MonetizationTagManager).singleton(),
     contentScript: asClass(ContentScript)
       .singleton()
       .inject(() => ({
