+++ src/content/lib/messages.ts
@@ -26,3 +26,21 @@ export const startMonetization = async (
     payload
   })
 }
+
+export const stopMonetization = async (
+  payload: ContentToBackgroundActionPayload[ContentToBackgroundAction.STOP_MONETIZATION]
+) => {
+  return await message.send({
+    action: ContentToBackgroundAction.STOP_MONETIZATION,
+    payload
+  })
+}
+
+export const resumeMonetization = async (
+  payload: ContentToBackgroundActionPayload[ContentToBackgroundAction.RESUME_MONETIZATION]
+) => {
+  return await message.send({
+    action: ContentToBackgroundAction.RESUME_MONETIZATION,
+    payload
+  })
+}
