--- src/content/monetizationTagManager/CustomError.ts
@@ -1,16 +1,0 @@
-export class CustomError extends Error {
-  constructor(message?: string) {
-    // 'Error' breaks prototype chain here
-    super(message)
-
-    // restore prototype chain
-    const actualProto = new.target.prototype
-
-    if (Object.setPrototypeOf) {
-      Object.setPrototypeOf(this, actualProto)
-    } else {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      ;(this as any).__proto__ = actualProto
-    }
-  }
-}
