+++ src/content/services/contentScript.ts
@@ -1,9 +1,15 @@
 import { Logger } from '@/shared/logger'
+import { MonetizationTagManager } from './monetizationTagManager'
 
 export class ContentScript {
-  constructor(private logger: Logger) {}
+  constructor(
+    private logger: Logger,
+    private monetizationTagManager: MonetizationTagManager
+  ) {}
 
   start() {
     this.logger.info('Content script started')
+
+    this.monetizationTagManager.start()
   }
 }
