+++ src/content/types.ts
@@ -5,10 +5,8 @@ export type MonetizationTagList = NodeListOf<MonetizationTag>
 
 export type MonetizationTagDetails = {
   walletAddress: WalletAddress | null
-  started: boolean
-  paused: boolean
-  stopped: boolean
-  //   attrs: getTagAttrs(tag, tagType), ?????? what for
-  //   initiatingUrl: this.window.location.href,  ?????? what for
-  //   fromBody: tag.parentElement != this.document.head ?????? what for
+  // started: boolean
+  // paused: boolean
+  // stopped: boolean
+  requestId: string
 }
