+++ src/shared/messages.ts
@@ -57,7 +57,9 @@ export type PopupToBackgroundMessage = {
 
 export enum ContentToBackgroundAction {
   CHECK_WALLET_ADDRESS_URL = 'CHECK_WALLET_ADDRESS_URL',
-  START_MONETIZATION = 'START_MONETIZATION'
+  START_MONETIZATION = 'START_MONETIZATION',
+  STOP_MONETIZATION = 'STOP_MONETIZATION',
+  RESUME_MONETIZATION = 'RESUME_MONETIZATION'
 }
 
 export interface CheckWalletAddressUrlPayload {
@@ -69,9 +71,19 @@ export interface StartMonetizationPayload {
   requestId: string
 }
 
+export interface StopMonetizationPayload {
+  requestId: string
+}
+
+export interface ResumeMonetizationPayload {
+  requestId: string
+}
+
 export interface ContentToBackgroundActionPayload {
   [ContentToBackgroundAction.CHECK_WALLET_ADDRESS_URL]: CheckWalletAddressUrlPayload
   [ContentToBackgroundAction.START_MONETIZATION]: StartMonetizationPayload
+  [ContentToBackgroundAction.STOP_MONETIZATION]: StopMonetizationPayload
+  [ContentToBackgroundAction.RESUME_MONETIZATION]: ResumeMonetizationPayload
 }
 
 export type ContentToBackgroundMessage = {
